Renamed species-selector-selection-modal-view to species-selector-results-view
