+++ src/content/app/new-species-selector/components/species-search-field/SpeciesSearchField.scss
@@ -10,6 +10,7 @@
 
 .speciesSearchField label {
   grid-area: label;
+  margin-left: 20px;
 }
 
 .input {
